backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
aiofiles
openai
PyJWT
bcrypt
httpx
